Add the initial prototype project that we set up in our first working session.
This is an Eclipse project that works with RepreZen API Studio Desktop (our Eclipse RCP disto) or with API Studio for Eclipse, available on the Eclipse Marketplace for installation into a standard Eclipse IDE, Oxygen or later. The core API specification in the /models folder is in standard OpenAPI 3.0 format. Code Generators in the GenTargets folder are a representative sample, and may change. These use OpenAPI-Tools/OpenAPI-Generator, through the RepreZen/GenFlow framework. More info to follow in the README.

Signed-off-by: Ted Epstein <ted.epstein@reprezen.com>